Collect stop-clock logic.
